requirements.txt
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==10.4.0
pycryptodome==3.23.0
pypng==0.20220715.0
qrcode==7.4.2
requests==2.32.4
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3

numpy
spidev